requirements.txt
selenium==4.1.0
ipykernel==6.7.0
python-decouple==3.5